refactor dir struct
